refactor(db): update schema for scope unification phase 3
- Remove scopes columns: type, ownerId, createdBy, timezone, notifyEmail, notifySlack
- Remove scopeTypeEnum
- Make scopes.clerkOrgId NOT NULL
- Make userId NOT NULL on secrets, variables, connectors, model_providers
- Make agent_runs.scopeId NOT NULL
- Update unique indexes to include userId for per-user resource isolation
- Delete org_access_tokens schema
