+++ turbo/apps/web/src/db/schema/agent-run.ts
@@ -23,7 +23,9 @@ export const agentRuns = pgTable(
   {
     id: uuid("id").defaultRandom().primaryKey(),
     userId: text("user_id").notNull(), // Clerk user ID - owner of this run
-    scopeId: uuid("scope_id").references(() => scopes.id), // Scope this run belongs to (nullable for backfill)
+    scopeId: uuid("scope_id")
+      .references(() => scopes.id)
+      .notNull(),
     agentComposeVersionId: varchar("agent_compose_version_id", {
       length: 64,
     }).references(() => agentComposeVersions.id, { onDelete: "set null" }),
